Quick prototype for DNS & HTTP
The DNS server, when queried with a TXT for a domain name of the form
covid19.pubhelth.{STATE}.us.gaia20.com responds with the URL of the
dashboard of STATE. STATE can be an abbreviation, such as `co`, or
`colorado`.

When the DNS server, is queried for A records, it responds with its own
IP, in hopes that the client will issue an HTTP request.

When the HTTP server receives a `GET http://covid19.pubhelth.{STATE}.us.gaia20.com/`,
it responds with a permanent redirect to the corresponding dashboard.